requirements.txt
click
pyyaml
python-frontmatter
setuptools